# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -5,12 +5,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
-import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.director.DirectorDao;
-import ru.yandex.practicum.filmorate.storage.feed.EventDao;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.genre.GenreDao;
 import ru.yandex.practicum.filmorate.storage.like.LikeDao;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
@@ -24,18 +21,14 @@ public class FilmService {
     private final UserStorage userStorage;
     private final LikeDao likeDao;
     private final DirectorDao directorDao;
-    private final GenreDao genreDao;
-    private final EventDao eventDao;
+
 
     @Autowired
-    public FilmService(@Qualifier("filmBdStorage") FilmStorage filmStorage, UserStorage userStorage, LikeDao likeDao,
-                       DirectorDao directorDao, GenreDao genreDao, EventDao eventDao) {
+    public FilmService(@Qualifier("filmBdStorage") FilmStorage filmStorage, UserStorage userStorage, LikeDao likeDao, DirectorDao directorDao) {
         this.filmStorage = filmStorage;
         this.userStorage = userStorage;
         this.likeDao = likeDao;
         this.directorDao = directorDao;
-        this.genreDao = genreDao;
-        this.eventDao = eventDao;
     }
 
     public List<Film> getFilms() {
@@ -64,34 +57,25 @@ public class FilmService {
 
     public void addLike(Long id, Long userId) {
         Film film = filmStorage.getFilmById(id);
-
-        likeDao.addLike(id, userId);
-        eventDao.addEvent(new Event(userId, id, "LIKE", "ADD"));
-        log.info("Добавлен лайк: фильм {}, пользователь id={}", film.getName(), userId);
+        if (film != null) {
+            likeDao.addLike(id, userId);
+            log.info("Добавлен лайк: фильм {}, пользователь id={}", film.getName(), userId);
+        }
     }
 
     public void deleteLike(Long id, Long userId) {
         Film film = filmStorage.getFilmById(id);
-
-        if (film.getLikes().contains(userId)) {
-            likeDao.deleteLike(id, userId);
-            eventDao.addEvent(new Event(userId, id, "LIKE", "REMOVE"));
-            log.info("Удален лайк: фильм {}, пользователь id={}", film.getName(), userId);
-        } else {
-            throw new NotFoundException("Лайк от указанного пользователя не найден");
+        if (film != null) {
+            if (film.getLikes().contains(userId)) {
+                likeDao.deleteLike(id, userId);
+                log.info("Удален лайк: фильм {}, пользователь id={}", film.getName(), userId);
+            } else {
+                throw new NotFoundException("Лайк от указанного пользователя не найден");
+            }
         }
     }
 
     public List<Film> getPopularFilm(int count, Integer genreId, Integer year) {
-        // проверим валидность присланного жанра
-        if (genreId != null) {
-            genreDao.findById(genreId);
-        } else
-            throw new NotFoundException("Попробуйте задать другой жанр для фильтрации популярных фильмов");
-        // проверим валидность присланного года
-        if (year != null && year < 0) {
-            throw new NotFoundException("Год не может быть отрицательным");
-        }
         return filmStorage.getPopularFilm(count, genreId, year);
     }
